+++ packages/web3-eth/test/unit/rpc_method_wrappers/fixtures/createAccessList.ts
@@ -51,6 +51,11 @@ export const testData: TestData[] = [
 		`${JSON.stringify(transaction)}\nblockNumber = BlockTags.PENDING`,
 		[transaction, BlockTags.PENDING],
 	],
+	[`${JSON.stringify(transaction)}\nblockNumber = BlockTags.SAFE`, [transaction, BlockTags.SAFE]],
+	[
+		`${JSON.stringify(transaction)}\nblockNumber = BlockTags.FINALIZED`,
+		[transaction, BlockTags.FINALIZED],
+	],
 	// blockNumber = Numbers
 	[`${JSON.stringify(transaction)}\nblockNumber = "0x4b7"`, [transaction, '0x4b7']],
 	[`${JSON.stringify(transaction)}\nblockNumber = 1207`, [transaction, 1207]],