+++ packages/web3-rpc-methods/test/unit/eth_rpc_methods/fixtures/get_block_transaction_count_by_number.ts
@@ -27,5 +27,7 @@ export const testData: TestData[] = [
 	['blockNumber = BlockTags.LATEST', [BlockTags.LATEST]],
 	['blockNumber = BlockTags.EARLIEST', [BlockTags.EARLIEST]],
 	['blockNumber = BlockTags.PENDING', [BlockTags.PENDING]],
+	['blockNumber = BlockTags.SAFE', [BlockTags.SAFE]],
+	['blockNumber = BlockTags.FINALIZED', [BlockTags.FINALIZED]],
 	['blockNumber = "0x4b7"', ['0x4b7']],
 ];