+++ packages/web3-validator/CHANGELOG.md
@@ -74,3 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 -   Moved `TypedArray` to `web3-types` (was also duplicated at `web3-utils`) (#5771)
+
+### Added
+
+-   Added support of `safe` and `finalized` block tags in `isBlockTag` method (#5823)